requirements.txt
scapy==2.5.0
pyyaml==6.0.1
requests==2.31.0
